Fix: add express server in vision-analyze-example
