+++ examples/vision-analyze-example/.env-template
@@ -1,1 +1,1 @@
-VITE_OPENAI_API_KEY=
+OPENAI_API_KEY=
