+++ examples/vision-analyze-example/package.json
@@ -9,10 +9,13 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "server": "node --env-file=.env src/server.mjs"
   },
   "dependencies": {
     "@goto-lab/pufu-editor": "^0.9.9",
+    "cors": "^2.8.5",
+    "express": "^4.21.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
